+++ src/commands/mod.rs
@@ -7,6 +7,7 @@
 
 pub mod by;
 pub mod find_in;
+pub mod for_file;
 pub mod info;
 pub mod kill;
 pub mod list;
@@ -19,6 +20,7 @@ pub mod unstick;
 
 pub use by::ByCommand;
 pub use find_in::InCommand;
+pub use for_file::ForCommand;
 pub use info::InfoCommand;
 pub use kill::KillCommand;
 pub use list::ListCommand;